+++ src/components/SettingsIcon/index.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import Style from './style';
+
+const SettingsIcon = (): JSX.Element =>
+  <Style>
+    <svg width='261' height='261' viewBox='0 0 261 261' fill='none'>
+      <circle cx='130' cy='130' r='111' fill='#C4C4C4' />
+      <path
+        d='M100.623 0H159.448L171.833 86.3808H88.2383L100.623 0Z'
+        fill='#C4C4C4'
+      />
+      <path
+        d='M159.448 261L100.623 261L88.2382 174.619L171.833 174.619L159.448 261Z'
+        fill='#C4C4C4'
+      />
+      <path
+        d='M17.1164 59.3489L58.7124 17.7529L128.55 70.0763L69.4398 129.186L17.1164 59.3489Z'
+        fill='#C4C4C4'
+      />
+      <path
+        d='M243.267 202.308L201.671 243.904L131.834 191.58L190.944 132.47L243.267 202.308Z'
+        fill='#C4C4C4'
+      />
+      <path
+        d='M261 100.623L261 159.448L174.619 171.833L174.619 88.2385L261 100.623Z'
+        fill='#C4C4C4'
+      />
+      <path
+        // eslint-disable-next-line max-len
+        d='M1.62401e-06 159.448L9.33806e-06 100.623L86.3808 88.2385L86.3808 171.833L1.62401e-06 159.448Z'
+        fill='#C4C4C4'
+      />
+      <path
+        d='M201.671 17.0962L243.267 58.6922L190.944 128.53L131.834 69.4195L201.671 17.0962Z'
+        fill='#C4C4C4'
+      />
+      <path
+        d='M58.7124 243.247L17.1164 201.651L69.4398 131.814L128.55 190.924L58.7124 243.247Z'
+        fill='#C4C4C4'
+      />
+      <circle cx='130' cy='130' r='62' fill='#FFFFFF' />
+    </svg>
+  </Style>;
+
+export default SettingsIcon;