+++ src/components/SettingsIcon/style.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const Style = styled.div`
+  cursor: pointer;
+  position: absolute;
+  left: 25px;
+  top: 25px;
+
+  svg {
+    height: 40px;
+    width: 40px;
+  }
+`;
+
+export default Style;