+++ docs/changelog.md
@@ -18,6 +18,8 @@ Ends October 14, 2020
 - update serve scripts to run all in one terminal.
 - add spacebar event listener.
 - update HAND_FULL enum to NOTIFY_HAND_FULL.
+- fix React used before it was defined issue.
+- add SettingsIcon.
 
 ## Sprint 4
 Ends September 16, 2020